+++ c-fractals/src/fractals.c
@@ -1,16 +1,16 @@
 #include "fractals.h"
 
 
-float _Complex
-fractal_julia(float _Complex z, float _Complex c)
+double _Complex
+fractal_julia(double _Complex z, double _Complex c)
 {
     return z*z + c;
 }
 
 
-float _Complex (*fractal_get(enum Fractal frac))(float _Complex, float _Complex)
+double _Complex (*fractal_get(enum Fractal frac))(double _Complex, double _Complex)
 {
-    float _Complex (*fptr)(float _Complex, float _Complex) = &fractal_julia;
+    double _Complex (*fptr)(double _Complex, double _Complex) = &fractal_julia;
     switch (frac) 
     {
         case FRAC_JULIA:
