+++ c-fractals/tests/test_optimized.c
@@ -5,23 +5,23 @@ MU_TEST(test_threaded_result)
 {
     int num_threads = 2;
 
-    float c_real = -0.835f;
-	float c_imag = -0.2321f;
+    double c_real = -0.835f;
+	double c_imag = -0.2321f;
 
-    float _Complex c = c_real + c_imag*I;
-    float R = ceilf(cabs(c)) + 1;
+    double _Complex c = c_real + c_imag*I;
+    double R = ceilf(cabs(c)) + 1;
 
     int ROWS = 100;
     int COLS = 100;
 
-    float x_start = -R;
-    float x_end = R;
+    double x_start = -R;
+    double x_end = R;
 
-    float y_start = -R;
-    float y_end = R;
+    double y_start = -R;
+    double y_end = R;
 
-    float x_step = (x_end - x_start) / ROWS;
-    float y_step = (y_end - y_start) / COLS;
+    double x_step = (x_end - x_start) / ROWS;
+    double y_step = (y_end - y_start) / COLS;
 
     int max_iterations = 100;
 
@@ -43,8 +43,8 @@ MU_TEST(test_threaded_result)
     fractal_get_colors(hCmatrix_nc, &fp);
     fractal_get_colors_th(hCmatrix_th, &fp, num_threads);
 
-    float max_color_nc = fractal_get_max_color(hCmatrix_nc);
-    float max_color_th = fractal_get_max_color(hCmatrix_th);
+    double max_color_nc = fractal_get_max_color(hCmatrix_nc);
+    double max_color_th = fractal_get_max_color(hCmatrix_th);
 
     MU_CHECK(max_color_nc == max_color_th);
